chore: Re-arrange IDE entity selectors for split
