+++ app/client/src/pages/Editor/IDE/EditorTabs/JSTab.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import clsx from "classnames";
 import { useSelector } from "react-redux";
 
-import type { EntityItem } from "@appsmith/selectors/appIDESelectors";
+import type { EntityItem } from "@appsmith/selectors/entitiesSelector";
 import { getCurrentPageId } from "@appsmith/selectors/entitiesSelector";
 import { useActiveAction } from "@appsmith/pages/Editor/Explorer/hooks";
 import { jsCollectionIdURL } from "@appsmith/RouteBuilder";
