+++ app/client/src/pages/Editor/IDE/EditorTabs/QueryTab.tsx
@@ -2,10 +2,12 @@ import React, { useMemo } from "react";
 import clsx from "classnames";
 import { useSelector } from "react-redux";
 
-import type { EntityItem } from "@appsmith/selectors/appIDESelectors";
-import { getAction } from "@appsmith/selectors/entitiesSelector";
-import { getPlugins } from "@appsmith/selectors/entitiesSelector";
-import { getCurrentPageId } from "@appsmith/selectors/entitiesSelector";
+import type { EntityItem } from "@appsmith/selectors/entitiesSelector";
+import {
+  getAction,
+  getPlugins,
+  getCurrentPageId,
+} from "@appsmith/selectors/entitiesSelector";
 import { useActiveAction } from "@appsmith/pages/Editor/Explorer/hooks";
 import history, { NavigationMethod } from "utils/history";
 import { getActionConfig } from "pages/Editor/Explorer/Actions/helpers";
